chore: Move the sync big-segment status provider back to big_segments
BigSegmentStoreStatusProviderImpl originally lived in big_segments.py; the
async work moved it into big_segments_common to share it. Now that async has
its own provider, the sync one is single-sided again, so move it back.
big_segments_common now holds only the shared I/O-free helpers.
