+++ ldclient/impl/big_segments_common.py
@@ -1,8 +1,7 @@
 """
-I/O-free helpers for big-segments status tracking, plus the sync
-:class:`BigSegmentStoreStatusProviderImpl`. The helpers (`_hash_for_user_key`,
-`is_stale`, `EMPTY_MEMBERSHIP`) are shared by both the sync
-:mod:`ldclient.impl.big_segments` and the async
+This module holds the shared, I/O-free helpers for big-segments status
+tracking. The helpers (`_hash_for_user_key`, `is_stale`, `EMPTY_MEMBERSHIP`)
+are shared by both the sync :mod:`ldclient.impl.big_segments` and the async
 :mod:`ldclient.impl.async_big_segments` managers; nothing here touches the
 store or network.
 """
@@ -10,13 +9,6 @@ store or network.
 import base64
 import time
 from hashlib import sha256
-from typing import Callable, Optional
-
-from ldclient.impl.listeners import Listeners
-from ldclient.interfaces import (
-    BigSegmentStoreStatus,
-    BigSegmentStoreStatusProvider
-)
 
 # use EMPTY_MEMBERSHIP as a singleton whenever a membership query returns None; it's safe to reuse it
 # because we will never modify the membership properties after they're queried
@@ -29,36 +21,3 @@ def _hash_for_user_key(user_key: str) -> str:
 
 def is_stale(timestamp: int, stale_after_millis) -> bool:
     return (timestamp is None) or ((int(time.time() * 1000) - timestamp) >= stale_after_millis)
-
-
-class BigSegmentStoreStatusProviderImpl(BigSegmentStoreStatusProvider):
-    """
-    Default implementation of the BigSegmentStoreStatusProvider interface.
-
-    The real implementation of getting the status is in the big segment store manager - we pass in a lambda that
-    allows us to get the current status from that class. So this class provides a facade for that, and
-    also adds the listener mechanism.
-    """
-
-    def __init__(self, status_getter: Callable[[], BigSegmentStoreStatus]):
-        self.__status_getter = status_getter
-        self.__status_listeners = Listeners()
-        self.__last_status = None  # type: Optional[BigSegmentStoreStatus]
-
-    @property
-    def status(self) -> BigSegmentStoreStatus:
-        return self.__status_getter()
-
-    def add_listener(self, listener: Callable[[BigSegmentStoreStatus], None]) -> None:
-        self.__status_listeners.add(listener)
-
-    def remove_listener(self, listener: Callable[[BigSegmentStoreStatus], None]) -> None:
-        self.__status_listeners.remove(listener)
-
-    def _update_status(self, new_status: BigSegmentStoreStatus):
-        last = self.__last_status
-        if last is None:
-            self.__last_status = new_status
-        elif new_status.available != last.available or new_status.stale != last.stale:
-            self.__last_status = new_status
-            self.__status_listeners.notify(new_status)
